+++ lib/CarlAudio/src/WAVFile.hpp
@@ -13,8 +13,8 @@
  */
 class WAVFile {
   private:
-    uint8_t* samples; /**< the samples cached for this file */
-    int16_t number_samples; /**< the number of samples cached for this file */
+    int16_t* samples; /**< the samples cached for this file */
+    uint32_t number_samples; /**< the number of samples cached for this file */
 
   public:
     /**
@@ -29,7 +29,8 @@ class WAVFile {
      * @param position which sample to get
      * @return the sample
      */
-    uint8_t get_sample(int16_t position);
+    uint8_t get_sample(uint32_t position);
+    // int16_t get_sample2(uint32_t position);
     /**
      * @brief get the number of samples
      * @return the number of samples
